refactor(chat): extract MessageBody from MessageCard for reuse
Move the body+attachments+extension-annotation rendering block out of
MessageCard into a dedicated MessageBody component. MessageBody accepts
a `compact` prop (default false) for the right-rail pinned panel (Task 16):
compact mode applies line-clamp-3 on body text, smaller image thumbnails,
hides extension action buttons, and collapses PDF/downloads to a chip row.
Lightbox state remains in the caller via an onImageClick callback.
